+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Dynamische inventory icons (o.a. flyers)
 - Politie kan nu barriers plaatsen die het NPC verkeer tegenhouden (F1 -> Politie -> Plaats Barrier)
+- UI fontsize is aanpasbaar in config menu
 
 ### Changed